# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Contacts app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of contacts app"
  },
  "servers": [
    {
      "url": "https://nodejs-hw-mongodb-13-9b96.onrender.com",
      "description": "Production server"
    },
    {
      "url": "http://127.0.0.1:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
    "/contacts": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get all contacts",
        "operationId": "getContacts",
        "description": "Get all user contacts with optional sorting and pagination.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "description": "Page number for pagination.",
            "schema": {
              "type": "integer",
              "example": 1,
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "perPage",
            "description": "Number of contacts per page.",
            "schema": {
              "type": "integer",
              "example": 10,
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "description": "Field to sort by (e.g., `name`).",
            "schema": {
              "type": "string",
              "example": "name",
              "default": "name"
            }
          },
          {
            "in": "query",
            "name": "sortOrder",
            "description": "Sort order (`asc` for ascending, `desc` for descending).",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "example": "asc",
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved user contact list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found contacts!"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "data",
                        "page",
                        "perPage",
                        "totalItems",
                        "totalPages",
                        "hasPreviousPage",
                        "hasNextPage"
                      ],
                      "properties": {
                        "data": {
                          "type": "array",
                          "description": "List of contacts matching filters, sorted, and paginated.",
                          "items": {
                            "$ref": "#/components/schemas/contacts"
                          }
                        },
                        "page": {
                          "type": "number",
                          "description": "Current page number.",
                          "example": 1
                        },
                        "perPage": {
                          "type": "number",
                          "description": "Number of contacts per page.",
                          "example": 10
                        },
                        "totalItems": {
                          "type": "number",
                          "description": "Total number of contacts matching the filters.",
                          "example": 15
                        },
                        "totalPages": {
                          "type": "number",
                          "description": "Total number of pages.",
                          "example": 2
                        },
                        "hasPreviousPage": {
                          "type": "boolean",
                          "description": "Indicates if there is a previous page.",
                          "example": false
                        },
                        "hasNextPage": {
                          "type": "boolean",
                          "description": "Indicates if there is a next page.",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "tags": [
          "Contacts"
        ],
        "summary": "Create a new contact",
        "operationId": "createContact",
        "description": "Create a new contact with the provided information.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "phoneNumber"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Name of the contact",
                    "example": "John Doe"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "description": "Phone number of the contact",
                    "example": "+1234567890"
                  },
                  "email": {
                    "type": "string",
                    "description": "Email address of the contact",
                    "example": "john.doe@example.com"
                  },
                  "photo": {
                    "type": "string",
                    "format": "binary",
                    "description": "Contact's photo file"
                  },
                  "isFavourite": {
                    "type": "boolean",
                    "description": "Mark contact as favorite",
                    "example": true
                  },
                  "contactType": {
                    "type": "string",
                    "description": "Type of contact",
                    "enum": [
                      "work",
                      "home",
                      "personal"
                    ],
                    "example": "personal"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contact created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Contact created successfully."
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/contacts"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "contacts": {
        "type": "object",
        "required": [
          "_id",
          "name",
          "phoneNumber",
          "userId",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "Unique ID of the contact.",
            "example": "674c7ea11a75750f5f9b86d4"
          },
          "name": {
            "type": "string",
            "description": "Full name of the contact.",
            "example": "John Doe"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Contact's phone number.",
            "example": "+1234567890"
          },
          "email": {
            "type": "string",
            "description": "Contact's email address.",
            "example": "johndoe@example.com"
          },
          "userId": {
            "type": "string",
            "description": "Unique ID of the contact's owner.",
            "example": "674dc213e880c5a6c4e30b19"
          },
          "photo": {
            "type": "string",
            "description": "Contact's photo URL.",
            "example": "Link to the contact's photo"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the contact was created (ISO 8601).",
            "example": "2024-12-01T12:34:56Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the contact was last updated (ISO 8601).",
            "example": "2024-12-02T08:15:30Z"
          },
          "isFavourite": {
            "type": "boolean",
            "description": "Mark contact as favorite",
            "example": true
          },
          "contactType": {
            "type": "string",
            "description": "Type of contact",
            "enum": [
              "work",
              "home",
              "personal"
            ],
            "example": "personal"
          }
        }
      }
    },
    "responses": {
      "400": {
        "description": "Bad request due to invalid request body, incorrect contact ID format, or missing required session ID.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "description": "HTTP error status.",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "description": "General error name.",
                  "example": "BadRequestError"
                },
                "data": {
                  "type": "object",
                  "description": "Details about the error.",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Detailed error explanation.",
                      "oneOf": [
                        {
                          "example": "Invalid request body"
                        },
                        {
                          "example": "Invalid contact ID format"
                        },
                        {
                          "example": "Session ID is required"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "Unauthorized. Please provide a valid token."
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "error"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "description": "HTTP error status.",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "description": "General error name.",
                  "example": "Something went wrong"
                },
                "error": {
                  "type": "string",
                  "description": "Detailed explanation of the server-side error.",
                  "example": "Internal server error"
                }
              }
            }
          }
        }
      }
    }
  }
}